Fix. Move Provider to file, add java doc
